Rename some files , seperate content controllers and services
