+++ src/controllers/contents/index.ts
@@ -1,0 +1,2 @@
+export { default as getPaginatedContents } from './getPaginatedContents';
+export { default as matchUserContent } from './matchUserContent';
