+++ src/controllers/users/index.ts
@@ -1,12 +1,12 @@
-export { default as pendingUsers } from './pendingUsers';
+export { default as pendingUsers } from './listOfPendingUsers';
+export { default as approvedUser } from './listOfApprovedUsers';
+export { default as rejectedUsers } from './lestOfRejectedUsers';
+
 export { default as approveUser } from './approveUser';
-export { default as approvedUser } from './getApprovedUsers';
 export { default as rejectUser } from './rejectUser';
-export { default as rejectedUsers } from './getRejectedUsers';
+export { default as blockUser } from './blockUser';
+
 export { default as editProfile } from './editProfile';
 export { default as createUser } from './addUser';
-export { default as blockUser } from './blockUser';
-export { default as getPaginatedContents } from './getPaginatedContents';
-export { default as matchUserContent } from './matchUserContent';
 export { default as changePassword } from './changePassword';
 export { default as getUserStatistics } from './getUserStatistics';
