+++ src/services/user/index.ts
@@ -1,6 +1,5 @@
 export { getUserByEmail, getUserById, addUser } from './user';
 export { default as getUsersStatus } from './userStatus';
-export { default as getContents } from './getContents';
 export { default as getUsers } from './getUsers';
 export { default as matchUserContent } from './matchUserContent';
 export { default as getNumberOfContent } from './getContentReport';
